Added 1D and 2D interpolated sampling

+++ main.cpp
@@ -414,7 +414,7 @@ static void rasterize(Camera& camera, int width, int height)
 
 							uv.y = uv.y - (int)uv.y;
 
-							const Vector4<unsigned char>& texel = texture->sample<Vector4<unsigned char>>(uv.x * texture->width, uv.y * texture->height);
+							const Vector4<unsigned char>& texel = texture->sample<Vector4<unsigned char>>((unsigned int)(uv.x * texture->width), (unsigned int)(uv.y * texture->height));
 							framebuffer.sample<Vector4<unsigned char>>(x, y) = Vector4<unsigned char>{ texel.z, texel.y, texel.x, 255 };
 						}
 					}
